feat: track updaters (#6)
This change introduces updaters tracking for both React and React
Native, enabling better tracking and profiling of component updates
within both environments. This functionality is crucial for enhancing
performance monitoring and debugging workflows in React-based
applications.

Updaters tracking: Implemented a mechanism to track updates for
components in both React and React Native. This allows for improved
tracking of component lifecycle events, aiding performance analysis and
troubleshooting.

Forked version of react-test-renderer: Due to the specific
requirements of updaters tracking, the standard react-test-renderer is
not sufficient for this functionality. Therefore, a forked version of
react-test-renderer has been used, which is now published under the
package name @react-render-measurement-tool/profiling-renderer. This
custom version includes the necessary tracking capabilities.